+++ src/drawer/index.unit.tsx
@@ -6,13 +6,6 @@ const defaultProps = {
 }
 
 describe('Drawer', () => {
-  beforeEach(() => {
-    jest.spyOn(window, 'requestAnimationFrame').mockImplementation(cb => cb())
-  })
-
-  afterEach(() => {
-    window.requestAnimationFrame.mockRestore()
-  })
   it('Renders with a custom className', () => {
     const customClassName = 'custom-drawer'
     const wrapper = shallow(
@@ -25,35 +18,38 @@ describe('Drawer', () => {
 
   it('Should call `onOpen` and `scrollLock` after open', () => {
     const onOpen = jest.fn()
-    const onChange = jest.fn()
+    const onTransitionEnd = jest.fn()
     const wrapper = mount(
-      <Drawer {...defaultProps} onChange={onChange} onOpen={onOpen}>
+      <Drawer {...defaultProps} onTransitionEnd={onTransitionEnd} onOpen={onOpen}>
         body
       </Drawer>,
     )
     const scrollLock = jest.spyOn(wrapper.instance(), 'scrollLock')
-    wrapper.find('.scrollableContent').simulate('transitionEnd')
-    expect(onChange).toHaveBeenCalledWith(false)
+    expect(onOpen).toHaveBeenCalledTimes(0)
+    expect(scrollLock).toHaveBeenCalledTimes(0)
     wrapper.setProps({ open: true })
-    wrapper.find('.scrollableContent').simulate('transitionEnd')
-    expect(onChange).toHaveBeenCalledWith(false)
+    wrapper.find('.kirk-drawer-scrollableContent').simulate('transitionEnd')
+    expect(onTransitionEnd).toHaveBeenCalledWith(true)
     expect(onOpen).toHaveBeenCalledTimes(1)
     expect(scrollLock).toHaveBeenCalledTimes(1)
   })
 
   it('Should call `onClose` and `scrollUnlock` after close', () => {
     const onClose = jest.fn()
-    const onChange = jest.fn()
+    const onTransitionEnd = jest.fn()
     const wrapper = mount(
-      <Drawer {...defaultProps} onChange={onChange} onClose={onClose}>
+      <Drawer {...defaultProps} open onTransitionEnd={onTransitionEnd} onClose={onClose}>
         body
       </Drawer>,
     )
     const scrollUnlock = jest.spyOn(wrapper.instance(), 'scrollUnlock')
-    wrapper.instance().close()
-    wrapper.find('.scrollableContent').simulate('transitionEnd')
+    expect(onClose).toHaveBeenCalledTimes(0)
+    expect(scrollUnlock).toHaveBeenCalledTimes(0)
+    wrapper.setProps({ open: false })
+    wrapper.find('.kirk-drawer-scrollableContent').simulate('transitionEnd')
+    expect(onTransitionEnd).toHaveBeenCalledWith(false)
     expect(onClose).toHaveBeenCalledTimes(1)
-    expect(onChange).toHaveBeenCalledWith(false)
     expect(scrollUnlock).toHaveBeenCalledTimes(1)
+
   })
 })